requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
lxml==3.8.0
prettytable==0.7.2
PyYAML==3.12
requests==2.18.4
urllib3==1.22
